test: refactor tests directory structure
Refactored directory structure of tests to mirror the Python
module structure.
