+++ tests/__init__.py
@@ -1,3 +1,3 @@
-#  This file is part of the QuestionPy Server. (https://questionpy.org)
-#  The QuestionPy Server is free software released under terms of the MIT license. See LICENSE.md.
+#  This file is part of QuestionPy. (https://questionpy.org)
+#  QuestionPy is free software released under terms of the MIT license. See LICENSE.md.
 #  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
