+++ tests/questionpy_server/collector/test_indexer.py
