+++ tests/questionpy_server/worker/worker/test_thread.py
@@ -1,4 +1,3 @@
-"""Contains tests specific to the :class:`ThreadWorker`."""
 #  This file is part of the QuestionPy Server. (https://questionpy.org)
 #  The QuestionPy Server is free software released under terms of the MIT license. See LICENSE.md.
 #  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
